Sensors/MPU9250: fix type casting.

+++ src/Sensors/MPU9250/Driver.hpp
@@ -915,7 +915,7 @@ private:
         }
         else if (m_is_spi)
         {
-          uint8_t tx_data[2] = {reg | READ_FLAG, 0};
+          uint8_t tx_data[2] = {static_cast<uint8_t>(reg | READ_FLAG), 0};
           uint8_t rx_data[2] = {0, 0};
           m_spi->transfer(tx_data, rx_data, 2);
           value = rx_data[1];
@@ -934,7 +934,7 @@ private:
         }
         else if (m_is_spi)
         {
-          uint8_t tx_data[2] = {reg | READ_FLAG, 0};
+          uint8_t tx_data[2] = {static_cast<uint8_t>(reg | READ_FLAG), 0};
           std::vector<uint8_t> rx_data(count);
           m_spi->transfer(tx_data, (uint8_t*) rx_data[0], count);
           dest = &rx_data[1];
